// change // PersistentAccountDAO and PersistentTransactionsDAO extends SQLiteOpenHelper. PersistentExpenseManager has objects from each of these two classes. constructor of the PersistentExpenseManager gets the Context as an argument, and it is used in the constructors of the above two classes. when PersistentExpenseManager object is created in MainActivity, context is passed to the PersistentExpenseManager, so that is can be used inside the class to create the above mentioned two classes.
// explanation //
the SQLiteOpenHelper abstract class has the method to create the database. but the class itself is not a database. once on object of this class is created the class calls its function "onCreate" and creates the database. therefore we extend this class from the classes we want to create a database. when objects of these classes are created, the super method called and database is created.